+++ packages/create-twilio-function/src/create-twilio-function/create-files.js
@@ -29,6 +29,7 @@ const typescriptDeps = {
 };
 const javaScriptDevDeps = { 'twilio-run': versions.twilioRun };
 const typescriptDevDeps = {
+  '@twilio/runtime-handler': versions.twilioRuntimeHandler,
   'twilio-run': versions.twilioRun,
   typescript: versions.typescript,
   copyfiles: versions.copyfiles,