+++ packages/runtime-handler/jest.config.js
@@ -4,9 +4,9 @@ module.exports = {
   ...base,
   globals: {
     'ts-jest': {
-      tsConfig: './tsconfig.test.json',
+      tsconfig: './tsconfig.test.json',
     },
   },
-  name: 'serverless-api',
-  displayName: 'serverless-api',
+  name: 'runtime-handler',
+  displayName: 'runtime-handler',
 };