+++ packages/runtime-handler/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0-rc.3](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio/runtime-handler@1.0.2...@twilio/runtime-handler@1.1.0-rc.3) (2021-05-24)
+
+
+### Features
+
+* extract runtime-handler and lazyLoading ([#252](https://github.com/twilio-labs/serverless-toolkit/issues/252)) ([4b11e69](https://github.com/twilio-labs/serverless-toolkit/commit/4b11e693248e44a8c6db4a95cf90e79e00f7db08))
+
+
+
+
+
 ## [1.0.2](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio/runtime-handler@1.0.2-beta.3...@twilio/runtime-handler@1.0.2) (2021-05-19)
 
 **Note:** Version bump only for package @twilio/runtime-handler